[Refactor] User 패키지 Role 을 서브모듈 Role 로 변경
