+++ user/src/main/java/com/haot/user/application/dto/req/AdminUserUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.haot.user.application.dto.req;
 
-import com.haot.user.domain.model.enums.Role;
+import com.haot.user.submodule.role.Role;
 import lombok.Builder;
 
 @Builder
