+++ user/src/main/java/com/haot/user/application/dto/res/AdminUserGetResponse.java
@@ -1,6 +1,6 @@
 package com.haot.user.application.dto.res;
 
-import com.haot.user.domain.model.enums.Role;
+import com.haot.user.submodule.role.Role;
 
 public record AdminUserGetResponse(
     String userId,
