+++ user/src/main/java/com/haot/user/application/dto/res/UserGetMeResponse.java
@@ -1,7 +1,7 @@
 package com.haot.user.application.dto.res;
 
 import com.haot.user.domain.model.enums.Gender;
-import com.haot.user.domain.model.enums.Role;
+import com.haot.user.submodule.role.Role;
 import lombok.Builder;
 
 @Builder
