+++ user/src/main/java/com/haot/user/application/service/UserCRUDServiceImpl.java
@@ -6,8 +6,8 @@ import com.haot.user.common.exception.UserException;
 import com.haot.user.common.exception.ErrorCode;
 import com.haot.user.common.util.Argon2PasswordEncoder;
 import com.haot.user.domain.model.User;
-import com.haot.user.domain.model.enums.Role;
 import com.haot.user.infrastructure.repository.UserRepository;
+import com.haot.user.submodule.role.Role;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
