--- user/src/main/java/com/haot/user/domain/model/enums/Role.java
@@ -1,12 +1,0 @@
-package com.haot.user.domain.model.enums;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public enum Role {
-  ADMIN("관리자"),
-  HOST("숙소 관리자"),
-  USER("일반 사용자");
-
-  private final String description;
-}
